+++ database/migrations/2024_06_28_160459_create_elim_duas_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('elim_duas', function (Blueprint $table) {
             $table->id();
             $table->string('namaKelompok');
-            $table->integer('jumlahPoin');
+            $table->double('jumlahPoin');
             $table->timestamps();
         });
     }
