+++ database/migrations/2024_06_28_211445_create_finals_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('finals', function (Blueprint $table) {
             $table->id();
             $table->string('namaKelompok');
-            $table->integer('jumlahPoin');
+            $table->double('jumlahPoin');
             $table->timestamps();
         });
     }
